Add magic round quote function
When hitting a " key in org-mode buffers, it instead inserts an "open"
round quote... unless we already have an open quote, in which case, we
insert a closing quote.

Works with single quotes as well, but with the added feature that
contractions and possessives put a closing quote even without an opening
quote.